Various tweaks
- Fixed save to library discard button show console error
- removed temporary inputselect 2 component
