--- packages/components/src/components/InputSelect2/index.js
@@ -1,4 +1,0 @@
-export {
-	default as InputSelect2
-}
-from './InputSelect2.vue'
