+++ packages/components/src/components/index.ts
@@ -28,7 +28,6 @@ export * from './InputNumber'
 export * from './InputRadio'
 export * from './InputRange'
 export * from './InputSelect'
-export * from './InputSelect2'
 export * from './InputShapeDividers'
 export * from './InputTextAlign'
 export * from './InputTextShadow'
