+++ packages/editor/src/components/AddElementPopup/ElementsTab.vue
@@ -219,6 +219,7 @@ export default {
 			display: flex;
 			flex-direction: column;
 			min-height: 385px;
+
 			.znpb-wrapper-category {
 				align-items: center;
 				max-height: 321px;
@@ -229,7 +230,7 @@ export default {
 			}
 		}
 
-		.znpb-fancy-scrollbar {
+		.znpb-wrapper-category {
 			flex-grow: 1;
 			padding: 0 6px 0 10px;
 		}
@@ -262,8 +263,12 @@ export default {
 .znpb-add-elements__filter {
 	display: flex;
 	padding: 0 10px;
+	margin-bottom: 20px;
 
 	&-category {
+		flex-grow: 1;
+		flex-shrink: 0;
+		width: 130px;
 		margin-right: 10px;
 
 		input[type="text"][readonly] {
@@ -286,10 +291,6 @@ export default {
 	grid-template-columns: 1fr 1fr 1fr;
 
 	&__search-wrapper.zion-input {
-		// width: calc(100% - 20px);
-		// padding: 0 10px;
-		margin-bottom: 20px;
-		// margin-left: 10px;
 		background: transparent;
 	}
 }
